Refactor embedding news into Qdrant DB and generate training data
- Refactor the embedding process of news articles into Qdrant DB
- Update the script for generating training data using GPT-4o
- Add doc strings for the functions
