+++ modules/dataset_wrangling/scripts/embed_news_into_qdrant.py
@@ -12,12 +12,12 @@ import sys
 import json
 
 from argparse import ArgumentParser
-from functools import partial
+import multiprocessing
 
 
 from loguru import logger
 from tqdm import tqdm
-from qdrant_client import QdrantClient
+
 
 sys.path.append(os.path.join(os.path.dirname(__file__), ".."))
 from src.paths import RAW_NEWS_PATH
@@ -30,10 +30,25 @@ from src.vector_db_api import (
 
 QDRANT_COLLECTION_NAME = "alpaca_news"
 VECTOR_SIZE = 384
+qdrant_client = get_qdrant_client()
+qdrant_client = init_collection(
+    qdrant_client,
+    QDRANT_COLLECTION_NAME,
+    VECTOR_SIZE,
+)
 
 
 def load_news(from_date: str, to_date: str) -> Dict:
+    """
+    Load news data from a JSON file.
 
+    Args:
+    - from_date: str: Start date in the format 'YYYY-MM-DD'.
+    - to_date: str: End date in the format 'YYYY-MM-DD'.
+
+    Returns:
+    - Dict: News data.
+    """
     filename = f"news_{from_date}_{to_date}.json"
     if not os.path.isfile(RAW_NEWS_PATH / filename):
         logger.error(f"News file: {filename} not found!!")
@@ -46,10 +61,17 @@ def load_news(from_date: str, to_date: str) -> Dict:
 
 
 def process_and_push_document(
-    qdrant_client: QdrantClient,
-    collection_name: str,
     article: Dict,
 ) -> None:
+    """
+    Process and push a news article into Qdrant.
+
+    Args:
+    - article: Dict: A news article.
+
+    Returns:
+    - None
+    """
     # Parsing the doc
     logger.debug("Parsing a new article")
     document = parse_article(article)
@@ -63,51 +85,57 @@ def process_and_push_document(
     document = embed_document(document)
 
     # Push document to the qdrant collection
-    push_document_to_qdrant(document, qdrant_client, collection_name)
+    push_document_to_qdrant(document, qdrant_client, QDRANT_COLLECTION_NAME)
 
 
 def embed_news_into_qdrant(
     news_data: List[Dict],
-    qdrant_client: QdrantClient,
-    collection_name: str,
     num_processes: int,
 ) -> None:
+    """
+    Embed news data into Qdrant.
 
+    Args:
+    - news_data: List[Dict]: List of news articles.
+    - num_processes: int: Number of system processes.
+
+    Returns:
+    - None
+    """
+    logger.info("Number of system processes: {num_processes}")
     if num_processes == 1:
         for doc in news_data:
-            process_and_push_document(qdrant_client, collection_name, doc)
+            process_and_push_document(doc)
 
     else:
-
-        # Parallel Processing.
-        import multiprocessing
-
-        # Create a partial function to ensure that all the data is loaded in the same collection of the qdrant client.
-        process_and_push_document_partial = partial(
-            process_and_push_document,
-            qdrant_client,
-            collection_name,
-        )
         with multiprocessing.Pool(processes=num_processes) as pool:
-            tqdm(
-                pool.imap(process_and_push_document_partial, news_data),
-                total=len(news_data),
-                desc="Processing",
-                unit="news",
+            _ = list(
+                tqdm(
+                    pool.imap(process_and_push_document, news_data),
+                    total=len(news_data),
+                    desc="Processing",
+                    unit="news",
+                )
             )
 
 
 def main(from_date: str, to_date: str, num_processes: int) -> None:
+    """
+    Main function to embed news data into Qdrant.
+
+    Args:
+    - from_date: str: Start date in the format 'YYYY-MM-DD'.
+    - to_date: str: End date in the format 'YYYY-MM-DD'.
+    - num_processes: int: Number of system processes.
+
+    Returns:
+    - None
+    """
     data = load_news(from_date, to_date)
+    logger.info(f"Number of news articles: {len(data)}")
 
-    qdrant_client = get_qdrant_client()
-    qdrant_client = init_collection(
-        qdrant_client,
-        QDRANT_COLLECTION_NAME,
-        VECTOR_SIZE,
-    )
-
-    embed_news_into_qdrant(data, qdrant_client, QDRANT_COLLECTION_NAME, num_processes)
+    logger.info("Processing and embedding news data into Qdrant")
+    embed_news_into_qdrant(data, num_processes)
 
 
 if __name__ == "__main__":
@@ -121,7 +149,7 @@ if __name__ == "__main__":
     parser.add_argument(
         "--to_date",
         type=str,
-        default="2024-01-09",
+        default="2024-01-30",
         help="End date in the format 'YYYY-MM-DD'.",
     )
     parser.add_argument(
@@ -131,5 +159,9 @@ if __name__ == "__main__":
         help="Number of system processes.",
     )
     args = parser.parse_args()
+
+    # Set logger level to INFO
+    logger.remove()
+    logger.add(sys.stdout, level="INFO")
 
     main(args.from_date, args.to_date, args.num_processes)
